Update resize comment
